Introducing NuriaTest, a minimalist testing framework
NuriaTest is a minimalist testing framwork in its current form. Its
advantage over QtTest is support for having test-code right in the
production code - Or outside if that's wanted. The main purpose of
having the test-code next to the production code is to allow rapid
development of new features without the hassle of setting up new
QtTests outside.

The two main macros are NURIA_TEST and NURIA_TEST_MAIN - The latter
may or may not be necessary for your use-case.