Resource and ObjectWrapperResource: Generic asynchronous API for RPC
Resource offers a generic interface for remote procedure calls. Network
is planned to offer a bunch of RPC servers and clients in the future,
e.g. for JSON-RPC. ObjectWrapperResource is able to wrap any QObject
into a Resource, making it possible to easily expose pretty much any
object for RPC with a RPC server implementation.